Update mb_vsprintf method
Phalcon 4.2+ debug returns an error when a simple warning occurs.
The list function returns a warning when the array created by the preg_split does not correspond to the number of parameters.

On vanilla PHP, this does not cause any problem, because it assigns null to the unsetted parameters by default. Unfortunately, on a newer version of Phalcon, 4.2+, it catches the warning and returns an error, even when using @.

To avoid this error, simply assign the variables manually instead of using the PHP list function.

While this fix is not the prettiest or most advanced, it allows us to avoid modifying the rest of the method which does not cause any problem.